progress on function migration
